+++ agent/agent.go
@@ -4,6 +4,9 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
+	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"io"
 	"log"
 	"net"
@@ -28,10 +31,12 @@ var dialer = websocket.DefaultDialer
 
 // Agent tunnels remote port on a gateway server to local destination.
 type Agent struct {
-	gatewayURL  string
-	key         string
-	service     ports.NetPort
-	destination string
+	gatewayURL           string
+	key                  string
+	service              ports.NetPort
+	destination          string
+	uplinkBytesCounter   prometheus.Counter
+	downlinkBytesCounter prometheus.Counter
 }
 
 // Start starts tunneling agents with given configurations. The agents and
@@ -43,12 +48,30 @@ func Start(conf config.Agent, ctx context.Context) error {
 
 	tasks := taskch.New()
 
+	uplinkBytesCounterVec := promauto.NewCounterVec(prometheus.CounterOpts{
+		Name: "uplink_bytes",
+		Help: "",
+	}, []string{
+		"destination",
+	})
+	downlinkBytesCounterVec := promauto.NewCounterVec(prometheus.CounterOpts{
+		Name: "downlink_bytes",
+		Help: "",
+	}, []string{
+		"destination",
+	})
+
 	for _, forw := range conf.Forwards {
+		upstreamBytesCounter := uplinkBytesCounterVec.WithLabelValues(forw.Destination)
+		downstreamBytesCounter := downlinkBytesCounterVec.WithLabelValues(forw.Destination)
+
 		agent := Agent{
-			gatewayURL:  conf.GatewayURL,
-			key:         conf.AuthKey,
-			service:     forw.Port,
-			destination: forw.Destination,
+			gatewayURL:           conf.GatewayURL,
+			key:                  conf.AuthKey,
+			service:              forw.Port,
+			destination:          forw.Destination,
+			uplinkBytesCounter:   upstreamBytesCounter,
+			downlinkBytesCounter: downstreamBytesCounter,
 		}
 
 		tasks.Go(func() error {
@@ -64,6 +87,11 @@ func Start(conf config.Agent, ctx context.Context) error {
 			}
 		})
 	}
+
+	tasks.Go(func() error {
+		http.Handle("/metrics", promhttp.Handler())
+		return http.ListenAndServe(conf.MetricsAddress, nil)
+	})
 
 	return tasks.Wait()
 }
@@ -189,9 +217,11 @@ func (agent *Agent) tunnel(accept service.BinderAcceptMessage, ctx context.Conte
 				return err
 			}
 
-			if _, err := io.CopyBuffer(conn, r, buf); err != nil {
+			n, err := io.CopyBuffer(conn, r, buf)
+			if err != nil {
 				return err
 			}
+			agent.uplinkBytesCounter.Add(float64(n))
 		}
 	})
 
@@ -208,6 +238,7 @@ func (agent *Agent) tunnel(accept service.BinderAcceptMessage, ctx context.Conte
 			if err := ws.WriteMessage(websocket.BinaryMessage, buf[:n]); err != nil {
 				return err
 			}
+			agent.downlinkBytesCounter.Add(float64(n))
 		}
 	})