+++ server/server.go
@@ -2,9 +2,9 @@ package server
 
 import (
 	"fmt"
-
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
+	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/snsinfu/reverse-tunnel/config"
 )
 
@@ -28,6 +28,16 @@ func Start(conf config.Server) error {
 	if conf.TLSConf.KeyPath != "" {
 		return e.StartTLS(conf.ControlAddress, conf.TLSConf.CertPath, conf.TLSConf.KeyPath)
 	}
+
+	// Expose metrics via another port.
+	go func() {
+		metrics := echo.New()
+		metrics.GET("/metrics", echo.WrapHandler(promhttp.Handler()))
+		_ = metrics.Start(conf.MetricsAddress)
+		// Or, using plain http instead:
+		//http.Handle("/metrics", promhttp.Handler())
+		//_ = http.ListenAndServe(conf.MetricsPort, nil)
+	}()
 
 	return e.Start(conf.ControlAddress)
 }